Add online/offline state
* Rename accountlist
* Cleanup unused actions
