+++ .eslintrc
@@ -13,6 +13,7 @@
   },
   "extends": "airbnb",
   "rules": {
+    "quote-props": 0,
     "no-plusplus": 0,
     "arrow-parens": 0,
     "global-require": 0,
