+++ src/account/__tests__/accountlistReducers-test.js
@@ -1,12 +1,12 @@
 import { fromJS } from 'immutable';
 import { LOGIN_SUCCESS, LOGOUT, ACCOUNT_REMOVE } from '../../account/accountActions';
-import accountlistReducers from '../accountlistReducers';
+import accountReducers from '../accountReducers';
 
 test('on login, update initial account with auth information', () => {
   const prevState = fromJS([{
     realm: 'http://realm.com',
   }]);
-  const newState = accountlistReducers(prevState, {
+  const newState = accountReducers(prevState, {
     type: LOGIN_SUCCESS,
     apiKey: '123',
     email: 'johndoe@example.com',
@@ -26,7 +26,7 @@ test('on login, if account does not exist, add as first item', () => {
     email: 'one@example.com',
     realm: 'http://realm1.com',
   }]);
-  const newState = accountlistReducers(prevState, {
+  const newState = accountReducers(prevState, {
     type: LOGIN_SUCCESS,
     apiKey: '456',
     email: 'two@example.com',
@@ -54,7 +54,7 @@ test('on login, if account does exist, merge new data, move to top', () => {
     realm: 'http://realm2.com',
     email: 'two@example.com',
   }]);
-  const newState = accountlistReducers(prevState, {
+  const newState = accountReducers(prevState, {
     type: LOGIN_SUCCESS,
     apiKey: '789',
     realm: 'http://realm2.com',
@@ -82,7 +82,7 @@ test('on logout, remove apiKey from active account, keep other information intac
     realm: 'http://realm2.com',
     email: 'two@example.com',
   }]);
-  const newState = accountlistReducers(prevState, { type: LOGOUT });
+  const newState = accountReducers(prevState, { type: LOGOUT });
   const expectedState = fromJS([{
     apiKey: '',
     realm: 'http://realm1.com',
@@ -101,7 +101,7 @@ test('on account removal, delete item from list', () => {
     realm: 'http://realm1.com',
     email: 'one@example.com',
   }]);
-  const newState = accountlistReducers(prevState, { type: ACCOUNT_REMOVE, index: 0 });
+  const newState = accountReducers(prevState, { type: ACCOUNT_REMOVE, index: 0 });
   const expectedState = fromJS([]);
   expect(newState.toJS()).toEqual(expectedState.toJS());
 });
