+++ src/account/accountReducers.js
@@ -1,40 +1,61 @@
 import { fromJS } from 'immutable';
 
 import {
-  ACCOUNT_ADD_SUCCEEDED,
-  DEV_EMAILS_SUCCEEDED,
-  DEV_EMAILS_FAILED,
-} from './accountActions';
+  REALM_ADD,
+  SET_AUTH_TYPE,
+  LOGIN_SUCCESS,
+  LOGOUT,
+  ACCOUNT_REMOVE,
+} from '../account/accountActions';
 
-
-const initialState = fromJS({
-  activeBackend: null,
-  authBackends: [],
-  directUsers: [],
-  directAdmins: [],
-  errors: [],
-});
+const initialState = fromJS([]);
 
 export default (state = initialState, action) => {
   switch (action.type) {
-    case ACCOUNT_ADD_SUCCEEDED: {
-      return state.merge(fromJS({
-        authBackends: action.authBackends,
-        activeBackend: null,
-        directUsers: [],
-        directAdmins: [],
+    case REALM_ADD: {
+      const accountIndex = state.findIndex(account =>
+        account.get('realm') === action.realm
+      );
+
+      if (accountIndex !== -1) {
+        return state
+          .unshift(state.get(accountIndex))
+          .delete(accountIndex + 1);
+      }
+
+      return state.unshift(fromJS({
+        realm: action.realm,
       }));
     }
-    case DEV_EMAILS_FAILED:
-      return state.merge({
-        errors: action.errors,
-      });
-    case DEV_EMAILS_SUCCEEDED:
-      return state.merge({
-        activeBackend: 'dev',
-        directUsers: action.directUsers,
-        directAdmins: action.directAdmins,
-      });
+    case SET_AUTH_TYPE:
+      return state.setIn([0, 'authType'], action.authType);
+    case LOGIN_SUCCESS: {
+      const accountIndex = state.findIndex(account =>
+        account.get('realm') === action.realm &&
+        (!account.get('email') || account.get('email') === action.email)
+      );
+
+      const { type, ...newAccount } = action; // eslint-disable-line no-unused-vars
+
+      if (accountIndex === -1) {
+        return state.unshift(newAccount);
+      }
+
+      if (accountIndex === 0) {
+        return state.mergeIn([0], newAccount);
+      }
+
+      const mergedAccount = state.get(accountIndex).merge(fromJS(newAccount));
+      return state
+        .unshift(mergedAccount)
+        .delete(accountIndex + 1);
+    }
+    case LOGOUT:
+      return state
+        .setIn([0, 'apiKey'], '');
+    case ACCOUNT_REMOVE:
+      return state
+        .delete(action.index);
     default:
       return state;
   }
