+++ src/reducers.js
@@ -1,22 +1,20 @@
 import { combineReducers } from 'redux';
-import accountlist from './accountlist/accountlistReducers';
-import app from './account/appReducers';
+import account from './account/accountReducers';
+import app from './app/appReducers';
 import errors from './error/errorReducers';
 import events from './events/eventReducers';
 import nav from './nav/navReducers';
 import stream from './stream/streamReducers';
 import subscriptions from './realm/subscriptionsReducers';
-import user from './account/accountReducers';
 import userlist from './userlist/userListReducers';
 
 export default combineReducers({
-  accountlist,
+  account,
   app,
   errors,
   events,
   nav,
   stream,
   subscriptions,
-  user,
   userlist,
 });
