+++ src/start/RealmScreen.js
@@ -3,7 +3,7 @@ import { TextInput } from 'react-native';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
-import { getAuth } from '../accountlist/accountlistSelectors';
+import { getAuth } from '../account/accountSelectors';
 import { styles, Screen, ErrorMsg, Button } from '../common';
 import { getAuthBackends } from '../api';
 import config from '../config';
